Post-merge fixes

+++ client/sql_string.cc
@@ -403,7 +403,7 @@ bool String::append(const char *s,uint32 arg_length)
 
 bool String::append(const char *s)
 {
-  return append(s, (uint) strlen(s));
+  return append(s, strlen(s));
 }
